requirements.txt
annotated-types==0.6.0
anyio==4.3.0
atomicwrites==1.4.1
attrs==23.2.0
bcrypt==4.1.2
blinker==1.7.0
certifi==2024.2.2
chardet==4.0.0
click==8.1.7
colorama==0.4.6
dataclasses==0.6
deprecation==2.1.0
dnspython==2.6.1
dominate==2.9.1
email_validator==2.1.1
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
gotrue==2.4.2
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==2.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
packaging==24.0
pluggy==1.5.0
postgrest==0.16.3
py==1.11.0
pydantic==2.7.1
pydantic_core==2.18.2
pytest==6.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
realtime==1.0.4
requests==2.25.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
storage3==0.7.4
StrEnum==0.4.15
supabase==2.4.3
supafunc==0.4.5
toml==0.10.2
typing_extensions==4.11.0
urllib3==1.26.18
visitor==0.1.3
websockets==12.0
Werkzeug==3.0.2
WTForms==3.1.2
